remove tmanager, restructure into module folders
