+++ .gitignore
@@ -2,6 +2,5 @@ node_modules
 build
 secrets.*.js
 *.zip
-utils/email.js
 docx_as_base64.txt
 \.idea
