--- src/js/email.js
@@ -1,58 +1,0 @@
-import constants from './util/constants';
-import connection from './connection';
-
-export default {
-
-    getEmailSettings() {
-
-        return new Promise((resolve, reject) => {
-            const email_settings_string = window.localStorage.getItem(constants.SETTINGS_EMAIL);
-            if (!email_settings_string) {
-
-                reject("Email Zugangsdaten fehlen. Siehe Einstellungen");
-                return;
-
-            }
-            const email_settings = JSON.parse(email_settings_string);
-
-            if (!email_settings.user ||
-                !email_settings.password ||
-                !email_settings.host ||
-                !email_settings.port ||
-                typeof email_settings.tls === 'undefined') {
-
-                reject("Email Zugangsdaten fehlen. Siehe Einstellungen");
-                return;
-
-            }
-
-            resolve(email_settings);
-        });
-    },
-
-    fetchAllMails(firstname, lastname, from) {
-        return this.getEmailSettings()
-            .then(email_settings => connection.post(constants.SERVER_FETCH_ALL_MAILS, {
-                settings: email_settings,
-                firstname: firstname,
-                lastname: lastname,
-                from: from
-            }))
-            .then(response => response.json());
-    },
-
-    /**
-     * try to log in
-     * returns the response body { status, error, data }
-     */
-    testConnection() {
-        return this.getEmailSettings()
-            .then(email_settings => connection.post(
-                constants.SERVER_TEST_CONNECTION, {
-                settings: email_settings
-            }
-            ))
-            .then(response => response.json());
-    }
-
-};
