+++ src/js/main/background.js
