+++ src/js/main/review_email/mail_generator.js
