+++ src/options.html
@@ -7,71 +7,6 @@
 </head>
 
 <body class="col">
-
-  <!-- Mail Server -->
-  <div class="row space-between">
-    <h1>Mail Server</h1>
-    <span id="status" class="bad"></span>
-  </div>
-
-  <div class="row skewered">
-    <div>
-      <label for="email_user">E-Mail</label>
-    </div>
-    <div>
-      <input type="text" id="email_user" />
-    </div>
-  </div>
-
-  <div class="row skewered">
-    <div>
-      <label for="email_password">Passwort</label>
-    </div>
-    <div>
-      <input type="password" id="email_password" />
-    </div>
-  </div>
-
-  <div class="row skewered">
-    <div>
-      <label for="email_host">Server</label>
-    </div>
-    <div>
-      <input type="text" id="email_host" />
-    </div>
-  </div>
-
-  <div class="row skewered">
-    <div>
-      <label for="email_port">Port</label>
-    </div>
-    <div>
-      <input type="text" id="email_port" />
-    </div>
-  </div>
-
-  <div class="row skewered">
-    <div>
-      <input type="checkbox" id="email_tls" />
-    </div>
-    <div>
-      <label for="email_tls">TLS</label>
-    </div>
-  </div>
-
-  <div class="row bad" id="save_email_message"></div>
-
-  <div class="row skewered">
-    <div></div>
-    <div>
-      <button class="good" id="save">speichern</button>
-    </div>
-    <div>
-      <button class="bad" id="wipe">alles löschen</button>
-    </div>
-  </div>
-
-  <hr />
   <!-- Check-in Dokument -->
   <div class="row space-between">
     <h1>Check-in Dokument</h1>
