+++ src/popup.html
@@ -57,17 +57,11 @@
       </div>
     </div>
 
-    <div class="row" id="tmanager">
-      <select id="tmanager_season"></select>
-      <button id="tmanager_fill" class="neutral">TManager ausfüllen</button>
-    </div>
-
     <!-- buttons for fill actions -->
     <div class="row">
       <button id="meldeschein_fill" class="neutral">Meldeschein ausfüllen</button>
       <button id="wlan_voucher_fill" class="neutral">WLAN Voucher ausfüllen</button>
       <button id="checkin_download" class="neutral">Check-in Dokument</button>
-      <button id="process_emails" class="neutral">Chiemgaukarte Mails erstellen</button>
     </div>
     <!-- "Popup" if first name's gender is unknown -->
     <div class="col hide" id="firstname_gender">
